Move stats commands to new package
