+++ src/main/java/space/npstr/wolfia/commands/util/CommandsCommand.java
@@ -34,9 +34,6 @@ import space.npstr.wolfia.commands.ingame.ShootCommand;
 import space.npstr.wolfia.commands.ingame.UnvoteCommand;
 import space.npstr.wolfia.commands.ingame.VoteCommand;
 import space.npstr.wolfia.commands.ingame.VoteCountCommand;
-import space.npstr.wolfia.commands.stats.BotStatsCommand;
-import space.npstr.wolfia.commands.stats.GuildStatsCommand;
-import space.npstr.wolfia.commands.stats.UserStatsCommand;
 import space.npstr.wolfia.config.properties.WolfiaConfig;
 import space.npstr.wolfia.domain.Command;
 import space.npstr.wolfia.domain.settings.ChannelSettingsCommand;
@@ -44,6 +41,9 @@ import space.npstr.wolfia.domain.setup.InCommand;
 import space.npstr.wolfia.domain.setup.OutCommand;
 import space.npstr.wolfia.domain.setup.SetupCommand;
 import space.npstr.wolfia.domain.setup.StatusCommand;
+import space.npstr.wolfia.domain.stats.BotStatsCommand;
+import space.npstr.wolfia.domain.stats.GuildStatsCommand;
+import space.npstr.wolfia.domain.stats.UserStatsCommand;
 
 import javax.annotation.Nonnull;
 import java.util.List;
